Make continuous variables render properly
Continuous variable rendering logic didn't use the getRoot function and so wasn't rendering the lines at all. Fix the rendering logic, and make lines render after bars to render on top of them. Also, make tail arrows render after bars and lines, to render on top of them, and make the title render last to render on top of everything.

Update version to 2.1.2 due to patches.