Refactor the distribution class-functions into es6 classes with proper inheritance
Each distribution is now a class and inherits from either DiscreteDistribution or ContinuousDistribution, both of which in turn inherit from Distribution. All of these classes are in the same directory, as is the file that defines the parseVariable function.

This refactoring also fixes a bug where evaluating variable.lessThan(0) would give the same results ass variable.lessThan(1). The code is now significantly easier to maintain since the definitions for lessThan, greaterThan and between (in the case of discrete distributions) are all in one place, and are the same for each variable.

Increment version number to 2.3.0.